some more chunk class remapping and byte, short -> float,int,double conversion

+++ src/java/com/mentalfrostbyte/jello/module/impl/gui/jello/Cords.java
@@ -59,7 +59,7 @@ public class Cords extends Module {
                                 + mc.player.getPosition().getZ();
                         float var6 = 85;
                         int var7 = var1.method13960();
-                        short var8 = 150;
+                        float var8 = 150;
                         float var9 = (float) ResourceRegistry.JelloLightFont18.getStringWidth(var5);
                         float var10 = Math.min(1.0F, (float) var8 / var9);
                         if (this.field23756.getDirection() != Direction.FORWARDS) {
